// src/app/components/inventory/models/params.ts
export class Params {
    maCt?: string = '';
    Ngay_Ct1?: any = '';
    Ngay_Ct2?: any = '';
    status?: string = '';
    top?: number = 50;
    keyword?: string = '';
    Ma_DvCs?: string = '';
    Is_App?: boolean= true;
    Ngay_KK?: any = '';
    Ma_Kho?: string = '';
    QTY?: number = 0;
    Stt?: string = '';
    Ma_Vt?: string = '';
 }